feat: 🎸 update chain name
