+++ src/constants/mainnet/index.ts
@@ -5,7 +5,7 @@ import * as BSC from '../platform/BSC';
 import * as BASE from '../platform/base';
 import DefaultWhitelistMap from './tokenWhitelist.json';
 
-import { DEFAULT_CHAIN_ICON, SupportedChainId, SupportedELFChainId } from '../chain';
+import { DEFAULT_CHAIN_ICON, DEFAULT_CHAIN_NAME, SupportedChainId, SupportedELFChainId } from '../chain';
 import { NetworkType } from 'types';
 
 export type ChainConstantsType = typeof MAINNET | typeof AELF | typeof tDVV | typeof BSC | typeof BASE;
@@ -45,25 +45,7 @@ export const ACTIVE_CHAIN: any = {
 };
 export const NATIVE_TOKEN_LIST = ['WETH', 'WBNB'];
 
-export const CHAIN_NAME: { [chainId in SupportedChainId | SupportedELFChainId]: string } = {
-  [SupportedChainId.MAINNET]: 'Ethereum Mainnet',
-  [SupportedChainId.KOVAN]: 'Kovan',
-  [SupportedChainId.GORELI]: 'Goerli',
-  [SupportedChainId.BSC_MAINNET]: 'Binance Smart Chain Mainnet',
-  [SupportedChainId.BSC_TESTNET]: 'Binance Smart Chain Testnet',
-  [SupportedChainId.HECO_MAINNET]: 'HECO',
-  [SupportedChainId.HECO_TESTNET]: 'HECO Testnet',
-  [SupportedChainId.OEC_MAINNET]: 'OEC',
-  [SupportedChainId.OEC_TESTNET]: 'OEC Testnet',
-  [SupportedChainId.POLYGON_MAINNET]: 'Polygon',
-  [SupportedChainId.POLYGON_TESTNET]: 'Polygon Testnet',
-  [SupportedELFChainId.AELF]: 'MainChain AELF Mainnet',
-  [SupportedELFChainId.tDVV]: 'SideChain tDVV Mainnet',
-  [SupportedELFChainId.tDVW]: 'SideChain tDVW Mainnet',
-  [SupportedChainId.SEPOLIA]: 'Sepolia Testnet',
-  [SupportedChainId.BASE_SEPOLIA]: 'Base Sepolia',
-  [SupportedChainId.BASE]: 'Base Mainnet',
-};
+export const CHAIN_NAME: { [chainId in SupportedChainId | SupportedELFChainId]: string } = DEFAULT_CHAIN_NAME;
 
 export const CHAIN_ICON = DEFAULT_CHAIN_ICON;
 
@@ -122,12 +104,12 @@ export const CrossChainTimeList = [
   {
     fromChainId: SupportedChainId.MAINNET,
     toChainId: [SupportedELFChainId.AELF, SupportedELFChainId.tDVV, SupportedELFChainId.tDVW],
-    time: '20',
+    time: '40',
   },
   {
     fromChainId: [SupportedELFChainId.AELF, SupportedELFChainId.tDVV, SupportedELFChainId.tDVW],
     toChainId: [SupportedChainId.MAINNET],
-    time: '20',
+    time: '40',
   },
   {
     fromChainId: SupportedChainId.BSC_MAINNET,
