+++ src/constants/platform/AELF.ts
@@ -1,8 +1,8 @@
 export const CHAIN_INFO = {
   chainId: 'AELF',
   exploreUrl: 'https://explorer.aelf.io/',
-  // rpcUrl: 'https://aelf-public-node.aelf.io',
-  rpcUrl: 'https://explorer.aelf.io/chain',
+  rpcUrl: 'https://aelf-public-node.aelf.io',
+  // rpcUrl: 'https://explorer.aelf.io/chain',
 };
 
 export const TOKEN_CONTRACT = 'JRmBduh4nXWi1aXgdUsj5gJrzeZb2LxmrAbf7W99faZSvoAaE';
