+++ src/constants/testnet/index.ts
@@ -8,7 +8,7 @@ import * as BSC_TESTNET from '../platform/BSC_Test';
 import * as SEPOLIA from '../platform/sepolia';
 import * as BASE_SEPOLIA from '../platform/base_sepolia';
 import DefaultWhitelistMap from './tokenWhitelist.json';
-import { DEFAULT_CHAIN_ICON, SupportedChainId, SupportedELFChainId } from '../chain';
+import { DEFAULT_CHAIN_ICON, DEFAULT_CHAIN_NAME, SupportedChainId, SupportedELFChainId } from '../chain';
 import { NetworkType } from 'types';
 import { IconInfo } from 'types/misc';
 
@@ -62,23 +62,10 @@ export const ACTIVE_CHAIN: any = {
 export const NATIVE_TOKEN_LIST = ['WETH', 'WBNB'];
 
 export const CHAIN_NAME: { [chainId in SupportedChainId | SupportedELFChainId]: string } = {
-  [SupportedChainId.MAINNET]: 'Ethereum',
-  [SupportedChainId.KOVAN]: 'Kovan',
-  [SupportedChainId.GORELI]: 'Goerli',
-  [SupportedChainId.BSC_MAINNET]: 'BSC',
-  [SupportedChainId.BSC_TESTNET]: 'Binance Smart Chain Testnet',
-  [SupportedChainId.HECO_MAINNET]: 'HECO',
-  [SupportedChainId.HECO_TESTNET]: 'HECO Testnet',
-  [SupportedChainId.OEC_MAINNET]: 'OEC',
-  [SupportedChainId.OEC_TESTNET]: 'OEC Testnet',
-  [SupportedChainId.POLYGON_MAINNET]: 'Polygon',
-  [SupportedChainId.POLYGON_TESTNET]: 'Polygon Testnet',
+  ...DEFAULT_CHAIN_NAME,
   [SupportedELFChainId.AELF]: 'MainChain AELF Testnet',
   [SupportedELFChainId.tDVV]: 'SideChain tDVV Testnet',
   [SupportedELFChainId.tDVW]: 'SideChain tDVW Testnet',
-  [SupportedChainId.SEPOLIA]: 'Sepolia Testnet',
-  [SupportedChainId.BASE_SEPOLIA]: 'Base Sepolia',
-  [SupportedChainId.BASE]: 'Base Mainnet',
 };
 
 export const CHAIN_ICON: { [chainId in SupportedChainId | SupportedELFChainId]: IconInfo } = {
@@ -162,12 +149,12 @@ export const CrossChainTimeList = [
   {
     fromChainId: SupportedChainId.SEPOLIA,
     toChainId: [SupportedELFChainId.AELF, SupportedELFChainId.tDVV, SupportedELFChainId.tDVW],
-    time: '20',
+    time: '40',
   },
   {
     fromChainId: [SupportedELFChainId.AELF, SupportedELFChainId.tDVV, SupportedELFChainId.tDVW],
     toChainId: SupportedChainId.SEPOLIA,
-    time: '20',
+    time: '40',
   },
 ];
